update(bin): ignored .gitmodules

+++ .bin/install.sh
@@ -11,6 +11,7 @@ link_to_homedir() {
         [[ ${bf} == ".git" ]] && continue
         [[ ${bf} == ".gitignore" ]] && continue
         [[ ${bf} == ".gitattributes" ]] && continue 
+        [[ ${bf} == ".gitmodules" ]] && continue
         if [[ -L "${HOME}/${bf}" ]]; then # is symlink
             command rm -f "${HOME}/${bf}"
         fi
